Move `ConfigurationCacheReport` to `o.g.configurationcache.problems`
For better cohesion.
